requirements.txt
aiogram
aiohttp
aioredis
python-dotenv
twilio
requests
SQLAlchemy
asyncpg
gino